--- phpcbf.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-./vendor/bin/phpcbf --standard=src/phpcs.ruleset.xml --extensions=php --tab-width=4 -sp src
