+++ src/Example/Php74/Common/EntityRepository.php
@@ -6,7 +6,7 @@
 
 declare(strict_types=1);
 
-namespace PixelFederation\CodingStandards\Example\Common;
+namespace PixelFederation\CodingStandards\Example\Php74\Common;
 
 interface EntityRepository
 {
