+++ src/Example/Php82/Bar.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PixelFederation\CodingStandards\Example\Php82;
+
+final class Bar
+{
+    public function __construct(
+        public readonly int $width,
+        public readonly int $height,
+    ) {}
+}
